file_manager/challenge/backend/requirements.txt
flask
playwright
flask_cors
flask-session
requests
python-dotenv